+++ mip/src/main/java/br/edu/utfpr/cp/emater/midmipsystem/service/mip/MIPSampleService.java
@@ -5,6 +5,7 @@ import br.edu.utfpr.cp.emater.midmipsystem.entity.mip.MIPSample;
 import br.edu.utfpr.cp.emater.midmipsystem.entity.mip.Pest;
 import br.edu.utfpr.cp.emater.midmipsystem.entity.mip.PestDisease;
 import br.edu.utfpr.cp.emater.midmipsystem.entity.mip.PestNaturalPredator;
+import br.edu.utfpr.cp.emater.midmipsystem.entity.security.MIPUserPrincipal;
 import br.edu.utfpr.cp.emater.midmipsystem.entity.survey.Survey;
 import br.edu.utfpr.cp.emater.midmipsystem.exception.AnyPersistenceException;
 import br.edu.utfpr.cp.emater.midmipsystem.exception.EntityAlreadyExistsException;
@@ -18,6 +19,8 @@ import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -62,9 +65,9 @@ public class MIPSampleService {
     private int calculateDaysAfterEmergence(Date emergenceDate, Date sampleDate) {
 
         long diffInMillies = Math.abs(sampleDate.getTime() - emergenceDate.getTime());
-        
+
         var result = TimeUnit.DAYS.convert(diffInMillies, TimeUnit.MILLISECONDS);
-        
+
         return (int) (result + 1);
     }
 
@@ -75,6 +78,13 @@ public class MIPSampleService {
     public void delete(Long anId) throws EntityNotFoundException, EntityInUseException, AnyPersistenceException {
 
         var existentSample = mipSampleRepository.findById(anId).orElseThrow(EntityNotFoundException::new);
+
+        var loggedUser = ((MIPUserPrincipal) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getUser();
+        var createdByName = existentSample.getCreatedBy() != null ? existentSample.getCreatedBy().getUsername() : "none";
+
+        if (!loggedUser.getUsername().equalsIgnoreCase(createdByName)) {
+            throw new AccessDeniedException("Usuário não autorizado para essa exclusão!");
+        }
 
         try {
             mipSampleRepository.delete(existentSample);
